AI_Surveillance_System/requirements.txt
ultralytics
opencv-python-headless
opencv-python
tensorflow
flask
matplotlib
seaborn
scikit-learn
numpy
pandas
Flask-SocketIO
Waitress
python-dotenv
